Move src/api and part of src/aedoc into api-extractor-model folder
