+++ apps/api-extractor-model/src/api/mixins/ApiStaticMixin.ts
