+++ apps/api-extractor-model/src/api/model/ApiIndexSignature.ts
