+++ apps/api-extractor-model/src/api/model/ApiTypeAlias.ts
