+++ apps/api-extractor-model/src/api/model/ApiVariable.ts
